# perf: numpy indexer use partial sort at query time