--- a/tests/unit/executors/indexers/test_numpyindexer.py
+++ b/tests/unit/executors/indexers/test_numpyindexer.py
@@ -147,7 +147,7 @@ def test_scipy_indexer_known_big(batch_size, compress_level, test_metas):
         np.testing.assert_equal(indexer.query_by_id([10000, 15000]), vectors[[0, 5000]])
 
 
-@pytest.mark.parametrize('batch_size, num_docs, top_k', [(1, 10, 1), (1, 10, 10), (10, 1, 1), (10, 1000, 10)])
+@pytest.mark.parametrize('batch_size, num_docs, top_k', [(1, 10, 1), (1, 10, 10), (10, 1, 1), (10, 1000, 10), (10, 10, 100)])
 def test__find_smallest_distances(batch_size, num_docs, top_k):
     dist = np.random.uniform(size=(batch_size, num_docs))