Hi, I have recently read your great blog article on your numpy indexer (https://hanxiao.io/2020/09/21/Numpy-Tricks-and-A-Strong-Baseline-for-Vector-Index/) and saw that at query time you do a full sort on the distances to retrieve the top k candidates.

This PR suggests using numpy's partial sort implementation `np.argpartition` which runs in O(n) versus O(n * log(n)) for `np.argsort`. 

With this code change the query method does not return the top k candidates ordered like with full sort before, but it seems the system is not relying on the top k candidates being ordered, at least the docstring does not imply it.

This little example illustrates the difference, the speedup for top 10 retrieval from a 100k array is around 12 fold
```python
# ipython 
import numpy as np

top_k = 10
dist = np.random.rand(100_000)

%%timeit 
idx_sort = dist.argsort()[:top_k]
# 9.4 ms ± 169 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

%%timeit
idx_partition = dist.argpartition(kth=top_k)[:top_k]
# 779 µs ± 2.85 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

idx_sort
# array([50384,  7687, 23873, 29129, 38318,   710, 51471, 81669, 47933, 61278])

idx_partition
# array([51471,   710, 47933, 38318, 50384, 61278, 81669, 29129, 23873, 7687])
```

I also ran the benchmarking script mentioned in the blog post (https://gist.github.com/hanxiao/43cad33b60cadd34f45236993689e5d9), the speedup is from 2.07s per run to 1.82s. When increasing the number of indexed documents from 10k to 100k and reducing their dimension from 10k to 1k, the speedup is from 4.05s per run to 3.14s.

Would be great to get your feedback on this :)