@JoanFM 2020-10-15
Hey @clennan ,

Thank you very much for your proposal, unfortunately, we already thought about this and we decided not to go for this solution at this moment as it adds the requirement to have the Ranker enforce ordering with not so clear benefit.

You can see the conversation at https://github.com/jina-ai/jina/pull/964,

You are welcome to change the docstring to show in a more clear way this limitation.