@JoanFM 2020-10-15
Hey @clennan,

I repoen the issue since I think maybe with this benchmarking of yours we could see if we decide to go for this direction. Sorry for the rushed reaction!