@JoanFM 2020-10-15
> ah, did not see the previous discussion, that's 1:1 to what I did here 😅
> 
> As mentioned in #964, the speedup only becomes significant when the index is large. Is there a reason why the benchmark was only done on a 10k index? I guess in practice the indexed data will be much larger and any optimisations would be targeted at those use cases anyways.
> 
> I agree that a full sort on the top k documents can be added to return sorted documents from the query method, should still be computationally much less expensive than a full sort on entire 

Adding first this argpartition and then sorting the top k in real order is a good solution, but then please provide some benchmark to ensure that this actually improve performance and that the extra cost of selecting the top_k is compensated by the sorting complexity gain.