@JoanFM 2020-10-15
Another interesting point could be adding a mode parameter to the indexing to know if it is a `match` or `ranking` index, so that the user can choose depending on wether it has a ranker downstream or it heavily relies in semantic distance. Any thoughts @hanxiao @bhavsarpratik @nan-wang @deepankarm ?