@JoanFM 2020-10-19
> Good morning, I have added the "hybrid" solution where a partial sort is done first and then a full sort on the top k. Here are some benchmarks:
> 
> ![image](https://user-images.githubusercontent.com/19587525/96416126-e282a200-11ef-11eb-9ef2-08baeb1325af.png)
> ![image](https://user-images.githubusercontent.com/19587525/96415437-e235d700-11ee-11eb-80d5-9521586036ed.png)
> 
> The y-axis measures the time difference between full sort (previous implementation) and partial sort in milliseconds, i.e. it is positive if the suggested new implementation is faster. The significant speedups occur if the number of documents (x-axis) in the index, and thus the number of calculated distances, is greater than 1million.
> 
> ![image](https://user-images.githubusercontent.com/19587525/96415471-f4b01080-11ee-11eb-971b-5108165dc79d.png)
> ![image](https://user-images.githubusercontent.com/19587525/96416202-fd551680-11ef-11eb-8fc0-cbdd9e978f71.png)
> ![image](https://user-images.githubusercontent.com/19587525/96415517-009bd280-11ef-11eb-82ed-38a9ebb2a802.png)
> 
> If the number of documents is below 5k the new implementation can be slightly slower with small batch sizes. I would argue that this performance loss is outweighed by the performance gains for a larger index or batch sizes.
> 
> The benchmark script can be found here
> https://gist.github.com/clennan/052115edf63a160ba160d081460df1d9

good analysis, thanks!