@bhavsarpratik 2020-10-19
Thanks @clennan for this addition! LGTM 👍 
@JoanFM Do we want to handle `n < k` - happens when we do testing by indexing a few docs but give a bigger k by mistake.