@JoanFM 2020-10-19
> Thanks @clennan for this addition! LGTM 
> @JoanFM Do we want to handle `n < k` - happens when we do testing by indexing a few docs but give a bigger k by mistake.

Good point, we should definitely handle it, I think it is already handled by this, but we should definitely add a `test` covering this case as extra security!. (Not important in that case what branch it goes to)