@clennan 2020-10-19
> Thanks for your contribution. Unfortunately, as you didn't set up your local git config correctly, the author of your commit does not match with your Github account. This poses problems:
> 
> * Your contribution can not be credited, and will not show up in the list of contributors (as bad git config = anonymous)
> * We can not accept CLA for anonymous contribution.
> 
> Please note that this does not only affect your contribution to Jina, but also affects all your contribution to anything on Github. So better fix it, once for all.
> Here is what you can do:
> 
> ```
> git config user.name "Correct new name"
> git config user.email "correct@new.email" -> match with your Github registered email address
> ```
> 
> Then fix the commit message by
> 
> ```
> git commit --amend --author="Author Name <email@address.com>" --no-edit
> git push --force
> ```

Amended all three commits